chatmap-api/pyproject.toml
[project]
name = "chatmap-api"
version = "0.1.0"
description = "ChatMap API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hotosm-auth[fastapi]==0.2.5",
    "altgraph==0.17.4",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "apscheduler==3.11.0",
    "certifi==2025.7.14",
    "chatmap-py==0.0.9",
    "click==8.2.1",
    "ecdsa==0.19.1",
    "fastapi==0.116.1",
    "geoalchemy2>=0.18.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "macholib==1.16.3",
    "packaging==25.0",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.6.1",
    "pycryptodome==3.23.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyinstaller==6.14.2",
    "pyinstaller-hooks-contrib==2025.8",
    "python-jose==3.5.0",
    "redis==6.2.0",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.47.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "uvicorn==0.35.0",
]
